move settings js to its own file
