+++ src/settings.js
@@ -1,0 +1,14 @@
+import domReady from '@wordpress/dom-ready';
+import { createRoot } from '@wordpress/element';
+
+const SettingsPage = () => {
+    return <div>Placeholder for settings page</div>;
+};
+
+domReady( () => {
+    const root = createRoot(
+        document.getElementById( 'vokab-settings' )
+    );
+
+    root.render( <SettingsPage /> );
+} );
